requirements.txt
beautifulsoup4==4.4.1
bleach==1.4.2
chardet==2.3.0
decorator==4.0.6
Django==1.8.8
django-contrib-comments==1.6.2
django-debug-toolbar==1.4
django-embed-video==1.0.0
filebrowser-safe==0.4.0
future==0.15.2
grappelli-safe==0.4.1
html5lib==0.9999999
imageio==1.4
Mezzanine==4.0.1
moviepy==0.2.2.11
numpy==1.10.4
oauthlib==1.0.3
Pillow==3.1.0
psycopg2==2.6.1
pytz==2015.7
PyVimeo==0.3.2
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
sqlparse==0.1.18
tqdm==3.7.1
tzlocal==1.2
wheel==0.24.0
yolk==0.4.3
